R: consolidate test models in one file
