+++ src/System.CommandLine.Tests/Binding/TestModels.cs
@@ -1,0 +1,172 @@
+﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using System.Threading.Tasks;
+
+namespace System.CommandLine.Tests.Binding
+{
+    public class ClassWithSingleLetterCtorParameter
+    {
+        public ClassWithSingleLetterCtorParameter(int x, string y)
+        {
+            X = x;
+            Y = y;
+        }
+
+        public int X { get; }
+
+        public string Y { get; }
+    }
+
+    public class ClassWithSingleLetterProperty
+    {
+        public int X { get; set; }
+
+        public int Y { get; set; }
+    }
+
+    public class ClassWithMultiLetterCtorParameters
+    {
+        public ClassWithMultiLetterCtorParameters(
+            int intOption = 123,
+            string stringOption = "the default",
+            bool boolOption = false)
+        {
+            IntOption = intOption;
+            StringOption = stringOption;
+            BoolOption = boolOption;
+        }
+
+        public int IntOption { get; }
+        public string StringOption { get; }
+        public bool BoolOption { get; }
+    }
+
+    public class ClassWithMultiLetterSetters
+    {
+        public int IntOption { get; set; }
+        public string StringOption { get; set; }
+        public bool BoolOption { get; set; }
+    }
+
+    public class ClassWithSettersAndCtorParametersWithDifferentNames
+    {
+        public ClassWithSettersAndCtorParametersWithDifferentNames(
+            int i = 123,
+            string s = "the default",
+            bool b = false)
+        {
+            IntOption = i;
+            StringOption = s;
+            BoolOption = b;
+        }
+
+        public int IntOption { get; set; }
+        public string StringOption { get; set; }
+        public bool BoolOption { get; set; }
+    }
+
+    public class ClassWithSettersAndCtorParametersWithMatchingNames
+    {
+        public ClassWithSettersAndCtorParametersWithMatchingNames(
+            int intOption = 123,
+            string stringOption = "the default",
+            bool boolOption = false)
+        {
+            IntOption = intOption;
+            StringOption = stringOption;
+            BoolOption = boolOption;
+        }
+
+        public int IntOption { get; set; }
+        public string StringOption { get; set; }
+        public bool BoolOption { get; set; }
+    }
+
+    public class ClassWithCtorParameter<T>
+    {
+        public ClassWithCtorParameter(T value) => Value = value;
+
+        public T Value { get; }
+    }
+
+    public class ClassWithSetter<T>
+    {
+        public T Value { get; set; }
+    }
+
+    internal class TypeWithInvokeAndCtor
+    {
+        public int IntValueFromCtor { get; }
+        public string StringValueFromCtor { get; }
+
+        public TypeWithInvokeAndCtor(int intFromCtor, string stringFromCtor)
+        {
+            IntValueFromCtor = intFromCtor;
+            StringValueFromCtor = stringFromCtor;
+        }
+
+        public int IntProperty { get; set; }
+        public string StringProperty { get; set; }
+
+        public Task<int> Invoke(string stringParam, int intParam)
+        {
+            return Task.FromResult(76);
+        }
+    }
+
+    internal class TypeWithInvokeNoCtor
+    {
+        public int IntProperty { get; set; }
+        public string StringProperty { get; set; }
+
+        public Task<int> Invoke(string stringParam, int intParam)
+        {
+            return Task.FromResult(66);
+        }
+
+        public Task<int> SomethingElse(int intParam, string stringParam)
+        {
+            return Task.FromResult(67);
+        }
+    }
+
+    internal class TypeWithStaticsInvokeAndCtor
+    {
+        public static int StaticIntProperty { get; set; } = 67;
+        public static string StaticStringProperty { get; set; }
+        public int IntValueFromCtor { get; }
+        public string StringValueFromCtor { get; }
+
+        public TypeWithStaticsInvokeAndCtor(int intFromCtor, string stringFromCtor)
+        {
+            IntValueFromCtor = intFromCtor;
+            StringValueFromCtor = stringFromCtor;
+        }
+
+        public int IntProperty { get; set; }
+        public string StringProperty { get; set; }
+
+        public static Task<int> Invoke(string stringParam, int intParam)
+        {
+            return Task.FromResult(96);
+        }
+    }
+
+    internal class TypeWithParameterlessInvokeAndCtor
+    {
+        // included for clarity
+        public TypeWithParameterlessInvokeAndCtor()
+        {
+        }
+
+        public int IntProperty { get; set; }
+        public string StringProperty { get; set; }
+
+        public Task<int> Invoke()
+        {
+            return Task.FromResult(86);
+        }
+    }
+}
