+++ src/System.CommandLine.Tests/Binding/TypeBinderTests.cs
@@ -283,95 +283,7 @@ namespace System.CommandLine.Tests.Binding
                 instance.StringOption.Should().Be("the default");
             }
         }
-
-        public class ClassWithSingleLetterCtorParameter
-        {
-            public ClassWithSingleLetterCtorParameter(int x, string y)
-            {
-                X = x;
-                Y = y;
-            }
-
-            public int X { get; }
-
-            public string Y { get; }
-        }
-
-        public class ClassWithSingleLetterProperty
-        {
-            public int X { get; set; }
-
-            public int Y { get; set; }
-        }
-
-        public class ClassWithMultiLetterCtorParameters
-        {
-            public ClassWithMultiLetterCtorParameters(
-                int intOption = 123,
-                string stringOption = "the default",
-                bool boolOption = false)
-            {
-                IntOption = intOption;
-                StringOption = stringOption;
-                BoolOption = boolOption;
-            }
-
-            public int IntOption { get; }
-            public string StringOption { get; }
-            public bool BoolOption { get; }
-        }
-
-        public class ClassWithMultiLetterSetters
-        {
-            public int IntOption { get; set; }
-            public string StringOption { get; set; }
-            public bool BoolOption { get; set; }
-        }
-
-        public class ClassWithSettersAndCtorParametersWithDifferentNames
-        {
-            public ClassWithSettersAndCtorParametersWithDifferentNames(
-                int i = 123,
-                string s = "the default",
-                bool b = false)
-            {
-                IntOption = i;
-                StringOption = s;
-                BoolOption = b;
-            }
-
-            public int IntOption { get; set; }
-            public string StringOption { get; set; }
-            public bool BoolOption { get; set; }
-        }
-
-        public class ClassWithSettersAndCtorParametersWithMatchingNames
-        {
-            public ClassWithSettersAndCtorParametersWithMatchingNames(
-                int intOption = 123,
-                string stringOption = "the default",
-                bool boolOption = false)
-            {
-                IntOption = intOption;
-                StringOption = stringOption;
-                BoolOption = boolOption;
-            }
-
-            public int IntOption { get; set; }
-            public string StringOption { get; set; }
-            public bool BoolOption { get; set; }
-        }
-
-        public class ClassWithCtorParameter<T>
-        {
-            public ClassWithCtorParameter(T value) => Value = value;
-
-            public T Value { get; }
-        }
-
-        public class ClassWithSetter<T>
-        {
-            public T Value { get; set; }
-        }
     }
+
+
 }
