--- src/System.CommandLine.Tests/Binding/TypeWithParameterlessInvokeAndCtor.cs
@@ -1,22 +1,0 @@
-﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System.Threading.Tasks;
-
-namespace System.CommandLine.Tests.Binding
-{
-    internal class TypeWithParameterlessInvokeAndCtor
-    {
-        // included for clarity
-        public TypeWithParameterlessInvokeAndCtor()
-        { }
-
-        public int IntProperty { get; set; }
-        public string StringProperty { get; set; }
-
-        public Task<int> Invoke()
-        {
-            return Task.FromResult(86);
-        }
-    }
-}
